rename hoem → board
