+++ .idea/gradle.xml
@@ -21,7 +21,7 @@
             <option value="$PROJECT_DIR$/features" />
             <option value="$PROJECT_DIR$/features/aboutapp" />
             <option value="$PROJECT_DIR$/features/auth" />
-            <option value="$PROJECT_DIR$/features/home" />
+            <option value="$PROJECT_DIR$/features/board" />
             <option value="$PROJECT_DIR$/features/notifications" />
             <option value="$PROJECT_DIR$/features/profile" />
             <option value="$PROJECT_DIR$/features/projects" />
