+++ app/build.gradle
@@ -39,7 +39,7 @@ android {
 }
 
 dependencies {
-    implementation project(':features:home')
+    implementation project(':features:board')
     implementation project(':features:notifications')
     implementation project(':features:profile')
     implementation project(':features:projects')
