+++ features/board/src/main/java/com/techcafe/todone/board/HomeFragment.kt
@@ -1,0 +1,12 @@
+package com.techcafe.todone.board
+
+import android.os.Bundle
+import android.view.View
+import androidx.fragment.app.Fragment
+
+class HomeFragment : Fragment(R.layout.fragment_home) {
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+    }
+}
