+++ features/board/src/main/res/drawable/ic_add_black_24dp.xml
