+++ features/board/src/main/res/values/strings.xml
