--- features/home/src/main/java/com/techcafe/todone/home/package-info.java
@@ -1,6 +1,0 @@
-@EpoxyDataBindingPattern(rClass = R.class,layoutPrefix = "model_")
-
-package com.techcafe.todone.home;
-
-import com.airbnb.epoxy.EpoxyDataBindingPattern;
-
