+++ features/settings/src/main/java/com/techcafe/todone/settings/SettingsFragment.kt
@@ -2,9 +2,7 @@ package com.techcafe.todone.settings
 
 import android.os.Bundle
 import android.view.View
-import androidx.fragment.app.Fragment
 import androidx.preference.PreferenceFragmentCompat
-import com.techcafe.todone.settings.databinding.FragmentSettingsBinding
 
 class SettingsFragment : PreferenceFragmentCompat() {
 
@@ -14,6 +12,5 @@ class SettingsFragment : PreferenceFragmentCompat() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-
     }
 }
