+++ settings.gradle
@@ -5,7 +5,7 @@ include ':corecomponent:androidcomponent'
 
 
 // Feature
-include ':features:home'
+include ':features:board'
 include ':features:notifications'
 include ':features:profile'
 // Data
